Don't make an archive unless the database exists.  Changes to be committed: 	modified:   CreateParameters.sh 	modified:   InstallPortableOpenERP.sh 	modified:   WipeClean.sh 	modified:   ipoerpMakeArchive.sh 	modified:   ipoerpMakeUpStartScript.sh 	new file:   ipoerpUnpackArchive.sh

+++ ipoerpMakeArchive.sh
@@ -2,8 +2,8 @@
 #
 DEFDIR=${0%/*}  #  Default directory of caller; maintains script portability.
 #
-# if [[  -z ${OOO} || -z ${SITENAME} || -z ${DATABASE_ARCHIVE}  || -z ${PSQLUSRDB}  ]]
-if [[                 -z ${SITENAME} || -z ${DATABASE_ARCHIVE}  || -z ${PSQLUSRDB}  ]]
+# if [[  -z ${OOO} || -z ${SITENAME} || -z ${DATABASE_ARCHIVE}  || -z ${PSQLUSRDB}  ||  -z  ${SITE_ARCHIVE}  ]]
+if [[                 -z ${SITENAME} || -z ${DATABASE_ARCHIVE}  || -z ${PSQLUSRDB}  ||  -z  ${SITE_ARCHIVE}  ]]
 then
 #
 echo "Usage :  ./ipoerpMakeArchive.sh"
@@ -11,12 +11,20 @@ echo "With required variables :"
 echo " -         SITENAME : ${SITENAME}"
 echo " - DATABASE_ARCHIVE : ${DATABASE_ARCHIVE}"
 echo " -        PSQLUSRDB : ${PSQLUSRDB}"
+echo " -     SITE_ARCHIVE : ${SITE_ARCHIVE}"
 exit 0
 fi
 
-su postgres -c "pg_dump ${PSQLUSRDB} | gzip -c > ${DATABASE_ARCHIVE}"
-#
-cp ${DATABASE_ARCHIVE} /srv/${SITENAME}/openerp/${PSQLUSRDB}.gz
-tar jcf ${SITE_ARCHIVE} /srv/${SITENAME}/openerp
+pushd /tmp
+if [[  $(su postgres -c "psql -l | grep -c '^ ${PSQLUSRDB}\b'") -gt 0 ]]
+then
+  su postgres -c "pg_dump ${PSQLUSRDB} | gzip -c > ${DATABASE_ARCHIVE}"
+  #
+  cp ${DATABASE_ARCHIVE} /srv/${SITENAME}/openerp/${PSQLUSRDB}.gz
+  tar jcf ${SITE_ARCHIVE} /srv/${SITENAME}/openerp
+else
+  echo "No database, so not worth making the archive."
+fi
+popd
 #
